refactor: 공모 목록 컴포넌트 분리
