+++ android/app/src/main/java/com/zzang/chongdae/presentation/view/home/component/HomeScreen.kt
@@ -3,10 +3,7 @@ package com.zzang.chongdae.presentation.view.home.component
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.foundation.layout.wrapContentWidth
@@ -15,12 +12,9 @@ import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.FloatingActionButton
-import androidx.compose.material3.FloatingActionButtonDefaults
 import androidx.compose.material3.HorizontalDivider
-import androidx.compose.material3.Icon
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextField
 import androidx.compose.material3.pulltorefresh.PullToRefreshBox
 import androidx.compose.material3.pulltorefresh.PullToRefreshDefaults.Indicator
 import androidx.compose.material3.pulltorefresh.rememberPullToRefreshState
@@ -29,7 +23,6 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.livedata.observeAsState
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -41,20 +34,16 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.Font
 import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.paging.LoadState
+import androidx.paging.compose.LazyPagingItems
 import androidx.paging.compose.collectAsLazyPagingItems
 import com.zzang.chongdae.R
-import com.zzang.chongdae.domain.model.Filter
-import com.zzang.chongdae.domain.model.FilterName
-import com.zzang.chongdae.domain.model.FilterType
+import com.zzang.chongdae.domain.model.Offering
 import com.zzang.chongdae.presentation.view.home.OfferingViewModel
 import com.zzang.chongdae.presentation.view.home.OnFloatingClickListener
 import com.zzang.chongdae.presentation.view.home.OnOfferingClickListener
-import kotlinx.coroutines.delay
 
-@OptIn(ExperimentalMaterial3Api::class)
 @Composable
 internal fun HomeScreen(
     viewModel: OfferingViewModel,
@@ -134,65 +123,7 @@ internal fun HomeScreen(
                 return@Surface
             }
 
-            val searchKeyword by viewModel.searchEvent.observeAsState()
-            val updatedOfferings by viewModel.updatedOffering.observeAsState(emptyList())
-            var isRefreshing by remember { mutableStateOf(false) }
-            val pullToRefreshState = rememberPullToRefreshState()
-
-            PullToRefreshBox(
-                isRefreshing = isRefreshing,
-                onRefresh = {
-                    isRefreshing = true
-                    viewModel.swipeRefresh()
-                    isRefreshing = false
-                },
-                state = pullToRefreshState,
-                indicator = {
-                    Indicator(
-                        modifier = Modifier.align(Alignment.TopCenter),
-                        isRefreshing = isRefreshing,
-                        containerColor = Color.White,
-                        color = colorResource(R.color.main_color),
-                        state = pullToRefreshState
-                    )
-                }
-            ) {
-                LazyColumn(
-                    Modifier
-                        .fillMaxSize()
-                        .padding(top = 15.dp)
-                        .padding(horizontal = 30.dp)
-                ) {
-                    items(count = offerings.itemCount) { index ->
-                        val item = offerings[index] ?: return@items
-                        val updatedItem = updatedOfferings.firstOrNull { it.id == item.id } ?: item
-                        OfferingItem(
-                            updatedItem,
-                            onOfferingClick,
-                            searchKeyword ?: "",
-                            R.color.search_keyword
-                        )
-                        if (index != offerings.itemCount - 1) {
-                            HorizontalDivider()
-                        }
-                    }
-                }
-
-                FloatingActionButton(
-                    onClick = { onFloatingClick.onClickFloatingButton() },
-                    shape = CircleShape,
-                    containerColor = colorResource(R.color.main_color),
-                    modifier = Modifier
-                        .padding(end = 20.dp, bottom = 30.dp)
-                        .align(Alignment.BottomEnd)
-                ) {
-                    Image(
-                        painter = painterResource(R.drawable.btn_main_create_offering),
-                        contentDescription = "Floating action button."
-                    )
-                }
-            }
+            OfferingsList(viewModel, offerings, onOfferingClick, onFloatingClick)
         }
     }
 }
-
